Split debian-emacs-bug.diff
After discussing with Reinhard, split debian-emacs-bug.diff in two
patches, one for the default Cc and one for the package version.
The former is not wanted in Ubuntu and causes extra work for each merge,
we'll apply it conditionally based on the system we're building one.
The latter is interesting no matter what, but remove the word 'Debian'
from the output.