Remove alternatives in reverse creation order (LP: #162932)
James Troup noticed that if the 'emacs' alternative is registered by
another package (e.g. emacs21), then removing the last subflavor leaves
it in manual state with a dangling link.

This is because we remove the subflavor alternative first, then try to
unregister from the 'emacs' alternative, but the latter points to the
former.  In that case update-alternatives switches the alternative to
manual mode because it thinks the administrator has manually removed the
subflavor link.

To fix this, remove the alternatives in the reverse order in which we
created them -- this way, the subflavor link is still here when we
unregister the 'emacs' alternative, so update-alternatives doesn't
switch it to manual mode.